DEV: Move out GithubPRPoster to its own file
